+++ Firmware/communication/interface_uart.h
@@ -9,11 +9,12 @@ extern "C" {
 #endif
 
 #include <cmsis_os.h>
+#include "usart.h"
 
 extern osThreadId uart_thread;
 extern const uint32_t stack_size_uart_thread;
 
-void start_uart_server(void);
+void start_uart_server(UART_HandleTypeDef* huart);
 void uart_poll(void);
 
 #ifdef __cplusplus