+++ tools/odrive/tests/can_test.py
@@ -109,9 +109,9 @@ class TestSimpleCAN():
 
     def run_test(self, odrive: ODriveComponent, canbus: CanInterfaceComponent, node_id: int, extended_id: bool, logger: Logger):
         odrive.disable_mappings()
-        odrive.handle.config.gpio15_mode = GPIO_MODE_CAN0
-        odrive.handle.config.gpio16_mode = GPIO_MODE_CAN0
-        odrive.handle.config.enable_can0 = True
+        odrive.handle.config.gpio15_mode = GPIO_MODE_CAN_A
+        odrive.handle.config.gpio16_mode = GPIO_MODE_CAN_A
+        odrive.handle.config.enable_can_a = True
         odrive.save_config_and_reboot()
 
         axis = odrive.handle.axis0